Error fixes & file renames
