+++ src/App.jsx
@@ -47,10 +47,10 @@ import GiftTreesComponet from "./pages/admin/gift/GiftTrees";
 import GCInventory from "./pages/admin/Inventory/GCInventory";
 import { DonationComponent } from "./pages/admin/donation/Donation";
 import CSRInventory from "./pages/admin/csr/CSRInventory";
-import { CampaignComponent } from "./pages/admin/campaign/campaign";
-import { CampaignsPage } from "./pages/admin/campaign/campaignpage";
-import { ReferralsPage } from "./pages/admin/campaign/referralpage"
-import { ReferralUserPage } from "./pages/admin/campaign/referraluser"
+import Campaigns from "./pages/admin/campaign/Campaign";
+import { CampaignsPage } from "./pages/admin/campaign/CampaignsPage";
+import { ReferralsPage } from "./pages/admin/campaign/ReferralsPage"
+import { ReferralUserPage } from "./pages/admin/campaign/ReferralUserPage"
 
 function App() {
   return (
@@ -106,7 +106,7 @@ function App() {
             <Route path="events" element={<EventsComponent />} />
             <Route path="site-inventory" element={<SiteInventory />} />
             <Route path="gc-inventory" element={<GCInventory />} />
-            <Route path="campaigns" element={<CampaignComponent />}></Route>
+            <Route path="campaigns" element={<Campaigns />}></Route>
             <Route path="tree-cards" element={<GiftTreesComponet />} />
             <Route path="donations" element={<DonationComponent />} />
             <Route path="corporate-dashboard" element={<CSRInventory />} />
